Change sensor bits to 16

+++ firmware/src/motor/motor.c
@@ -166,7 +166,7 @@ bool motor_calibrate_inductance(void)
     return true;
 }
 
-TM_RAMFUNC uint8_t motor_find_pole_pairs(uint16_t ticks, float mpos_start, float mpos_end, float epos_rad)
+TM_RAMFUNC uint8_t motor_find_pole_pairs(uint32_t ticks, float mpos_start, float mpos_end, float epos_rad)
 {
 	const float mpos_diff = our_fabsf(mpos_end - mpos_start);
 	float mpos_diff_rad = TWOPI * mpos_diff / ticks;
